requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-filter==22.1
Pillow==9.4.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
